% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grsofun_collect.R
\name{grsofun_collect}
\alias{grsofun_collect}
\title{Reads daily output and aggregates temporally. Returning the data is optional.
By default, the aggregated data is written to tidy files by longitudinal bands.}
\usage{
grsofun_collect(settings, return_data = FALSE)
}
\arguments{
\item{settings}{...}

\item{return_data}{Flag whether to return loaded data in the R session. If FALSE
only RDS files are written to disk, if TRUE RDS files are
written to disk AND the data.frame is returned.}
}
\description{
Reads daily output and aggregates temporally. Returning the data is optional.
By default, the aggregated data is written to tidy files by longitudinal bands.
}
