% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grsofun_tidy.R
\name{grsofun_tidy}
\alias{grsofun_tidy}
\title{Tidies necessary input data to prepare grsofun_run()}
\usage{
grsofun_tidy(settings, ...)
}
\arguments{
\item{...}{Additional arguments that are passed to map2tidy for ALL processed
input types. The documentation of the allowed arguments can be
accessed with \code{?map2tidy}}

\item{setting}{A list of grsofun settings.}
}
\value{
Returns a data.frame with the filenames of the rds files created and
a status message for each file.
}
\description{
Tidies necessary input data to prepare grsofun_run()
}
